fix: AI에너지 리포트 오류 수정 - AI에너지 리포트 오류 수정
Jira: feature/mobile/design-S13P21D208-319

+++ frontend/app/src/main/java/com/example/eeum/data/model/response/report/AiSummaryLastYearSameMonth.kt
@@ -2,5 +2,5 @@ package com.example.eeum.data.model.response.report
 
 data class AiSummaryLastYearSameMonth(
     val bill: Int,
-    val usage: Int
+    val usage: Double
 )
