+++ frontend/app/src/main/java/com/example/eeum/data/model/response/report/AiSummaryPeriod30d.kt
@@ -3,5 +3,5 @@ package com.example.eeum.data.model.response.report
 data class AiSummaryPeriod30d(
     val end: String,
     val start: String,
-    val totalKwh: Int
+    val totalKwh: Double
 )
